+++ packages/orchestrator-ui-components/src/components/forms/CreateForm.tsx
@@ -16,26 +16,13 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import { Form, FormNotCompleteResponse } from '../../types/forms';
 import UserInputFormWizard from './UserInputFormWizard';
-import { BaseApiClient } from '../../api';
+import {apiClient, BaseApiClient} from '../../api';
 
 interface IProps {
     preselectedInput?: any;
     formKey: string;
     handleSubmit: (userInputs: any) => void;
 }
-
-abstract class ApiClientInterface extends BaseApiClient {
-    abstract cimStartForm: (formKey: string, userInputs: {}[]) => Promise<any>;
-}
-class ApiClient extends ApiClientInterface {
-    cimStartForm = (
-        formKey: string,
-        userInputs: {}[],
-    ): Promise<{ id: string }> => {
-        return this.postPutJson(formKey, userInputs, 'post', false, true);
-    };
-}
-const apiClient: ApiClient = new ApiClient();
 
 export default function CreateForm(props: IProps) {
     const { preselectedInput, formKey, handleSubmit } = props;
@@ -44,23 +31,13 @@ export default function CreateForm(props: IProps) {
 
     const submit = useCallback(
         (userInputs: {}[]) => {
-            // if (preselectedInput.product) {
-            // Todo: decide if we want to implement pre-selections and design a way to do it generic
-            //     userInputs = [{ preselectedInput }, ...userInputs];
-            // }
-            let promise = apiClient.cimStartForm(formKey, userInputs).then(
+            return apiClient.cimStartForm(formKey, userInputs).then(
                 (form) => {
                     console.log('Submit {formkey} =', formKey);
                     handleSubmit(form);
                 },
-                (e) => {
-                    throw e;
-                },
-            );
 
-            return apiClient.catchErrorStatus<any>(promise, 503, (json) => {
-                alert('/tickets');
-            });
+            );
         },
         [formKey, handleSubmit],
     );
